requirements.txt
# requirements.txt
fastapi
uvicorn
python-multipart
requests
Pillow
pydub  # Para manejar archivos de audio
scipy
torch
transformers
bitsandbytes
auto-gptq
accelerate
